updates
updated example, now middlewares contained at main.go
